feat(regexp): replace regex with regexp
